Simplify our backup/restore procedure
We now have a the same situation at backup and restore time: a pwd to a
tmp folder with all items to backup/restore. This makes it much more
convenient to add our new feature: the possibility to execute arbitrary
commands at backup/restore time (such as dumping/restoring a DB).